% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Kmeans.R
\name{KMEANS}
\alias{KMEANS}
\title{K-means Clustering Algorithm}
\usage{
KMEANS(data.matrix, K, max_iters = 100)
}
\arguments{
\item{data.matrix}{The data matrix for the clustering}

\item{K}{The number of clusters}

\item{max_iters}{The maximum number of iterations for convergence}
}
\value{
A list containing the results of the K-means clustering algorithm
}
\description{
This function performs the K-means clustering algorithm on a given data matrix
}
\examples{
# Generate a test dataset
set.seed(123)
data.matrix <- matrix(rnorm(300), ncol = 2)
# Perform K-means clustering
K <- 3
result <- KMEANS(data.matrix, K)
# Print the cluster assignments and centers
print(result)


}
